[RSPEC] setup using of shared contexts
    **THIS PR COVERS:**

    1) RSpec: Adding of support of shared contexts

    2) Moved Sequence to separated file, so that I can re-use it in anothers samples

    3) Removed no longer needed '.gitkeep' files from 'samples' and 'spec' folders
