+++ spec/01_my_enumerable_spec.rb
@@ -1,20 +1,13 @@
-require "./samples/01_my_enumerable.rb"
+require "spec_helper.rb"
+require "./samples/sequence.rb"
 
 describe "My Enumerable Module" do
-  let(:sequence) do
-    #
-    # It will generate following sequence with 5 elements
-    # starting from 0 and with step 5:
-    #
-    # [0, 5, 10, 15, 20]
-    #
-    Sequence.new(0, 5, 5)
-  end
+  include_context "sequence"
 
   it "should provide ability to iterate sequence via #map method" do
     res = sequence.map { |x| x + 1 }
 
-    expect(res).to be_eql(
+    expect(res).to match_array(
       [1, 6, 11, 16, 21]
     )
   end
@@ -37,25 +30,25 @@ describe "My Enumerable Module" do
     #
     # 0..4
     #
-    expect(res).to be_eql(0..4)
+    expect(res).to match_array(0..4)
   end
 
   it "should provide ability to iterate sequence via #select method" do
     res = sequence.select { |x| x > 10 }
 
-    expect(res).to be_eql(
+    expect(res).to match_array(
       [15, 20]
     )
 
     res = sequence.select { |x| x.even? }
 
-    expect(res).to be_eql(
+    expect(res).to match_array(
       [0, 10, 20]
     )
 
     res = sequence.select { |x| x.odd? }
 
-    expect(res).to be_eql(
+    expect(res).to match_array(
       [5, 15]
     )
   end
