+++ spec/spec_helper.rb
@@ -1,0 +1,5 @@
+Dir["./spec/support/**/*.rb"].each { |f| require f }
+
+RSpec.configure do |config|
+  config.shared_context_metadata_behavior = :apply_to_host_groups
+end
