+++ spec/support/shared_contexts/sequence.rb
@@ -1,0 +1,11 @@
+shared_context "sequence" do
+  let(:sequence) do
+    #
+    # It will generate following sequence with 5 elements
+    # starting from 0 and with step 5:
+    #
+    # [0, 5, 10, 15, 20]
+    #
+    Sequence.new(0, 5, 5)
+  end
+end
